Change doubles to floats

+++ demos/mandelbrot/main.cpp
@@ -51,21 +51,21 @@ class MandelbrotApp
 #endif
 {
   // Use doubles for more zoom
-  MandelbrotCalculator<double> m_calc;
+  MandelbrotCalculator<float> m_calc;
 
   // Texture for displaying the set
   ci::gl::Texture2dRef m_tex;
 
   // Coordinates of the center point
-  double m_ctr_x = 0;
-  double m_ctr_y = 0;
+  float m_ctr_x = 0;
+  float m_ctr_y = 0;
 
   // The viewable range on Y axis
-  double m_range = 1;
+  float m_range = 1;
 
   // Mouse coordinates from previous click
-  double m_prev_mx = 0;
-  double m_prev_my = 0;
+  float m_prev_mx = 0;
+  float m_prev_my = 0;
 
  public:
   MandelbrotApp() : m_calc(WIDTH, HEIGHT) {}
@@ -80,13 +80,13 @@ class MandelbrotApp
   void update() override {
     // Transform coordinates from the ones used here - center point
     // and range - to the ones used in MandelbrotCalculator - min and max X, Y.
-    double range_x = m_range * double(WIDTH) / double(HEIGHT);
+    float range_x = m_range * float(WIDTH) / float(HEIGHT);
     auto half_x = range_x / 2.0f;
-    double min_x = m_ctr_x - half_x;
-    double max_x = m_ctr_x + half_x;
+    float min_x = m_ctr_x - half_x;
+    float max_x = m_ctr_x + half_x;
     auto half_y = m_range / 2.0f;
-    double min_y = m_ctr_y - half_y;
-    double max_y = m_ctr_y + half_y;
+    float min_y = m_ctr_y - half_y;
+    float max_y = m_ctr_y + half_y;
 
     // Set new coordinates and recalculate the fractal
     m_calc.set_bounds(min_x, max_x, min_y, max_y);
@@ -121,8 +121,8 @@ class MandelbrotApp
 
   void mouseDrag(ci::app::MouseEvent event) override {
     // Calculate normalized coordinates
-    auto x = event.getX() / double(WIDTH);
-    auto y = event.getY() / double(HEIGHT);
+    auto x = event.getX() / float(WIDTH);
+    auto y = event.getY() / float(HEIGHT);
 
     // Find the difference from last click
     auto dx = m_prev_mx - x;
@@ -132,12 +132,12 @@ class MandelbrotApp
     // If the difference is big enough, drag the center point
     // and with it the viewable part of the plane. The epsilon
     // is necessary to avoid noisy jumps
-    constexpr double EPS = .1f;
+    constexpr float EPS = .1f;
     if (dx < EPS && dx > -EPS) {
       m_ctr_x += dx * m_range;
     }
     if (dy < EPS && dy > -EPS) {
-      m_ctr_y += dy * m_range * double(WIDTH) / double(HEIGHT);
+      m_ctr_y += dy * m_range * float(WIDTH) / float(HEIGHT);
     }
 
     m_prev_mx = x;
